feat(codes/cpp): revert the changes `size_t` back to `int`

+++ codes/cpp/chapter_array_and_linkedlist/list.cpp
@@ -48,8 +48,8 @@ int main() {
     PrintUtil::printVector(list);
 
     /* 通过索引遍历列表 */
-    size_t count = 0;
-    for (size_t i = 0; i < list.size(); i++) {
+    int count = 0;
+    for (int i = 0; i < list.size(); i++) {
         count++;
     }
